+++ docs/changes.rst
@@ -13,6 +13,7 @@ Release 2.1.1
 Fixes
 
 * Fix infinite recursion in yaml_load when no yaml files present (ZPS-3880)
+* Update default value of threshold's escalationCount in the documentation (ZPS-4031)
 
 
 Release 2.1.0